+++ CHANGELOG.md
@@ -10,7 +10,6 @@ A new section Unreleased is opened then for next dev phase.
 - Add Contributor Licence Agreement. [#37]
 - Add cumulative Density Function stats [#34]
 - Add Nuth/Kaab iterations number option [#13]
-
 - Add option to give a local geoid model. [#33]
 
 ### Changed
@@ -18,6 +17,7 @@ A new section Unreleased is opened then for next dev phase.
 ### Fixed
 - Fix numpy, cython pip upgrade install [#37]
 - Fix remove_outliers no_data 0 bug. [#43]
+- Fix input images orientation possible bug [#49]
 
 
 ## 0.1.0 First Open Source Official Release (July 2020)